% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{chat_slack}
\alias{chat_slack}
\title{Interact with Chat API}
\usage{
chat_slack(
  slack_method = "chat.postMessage",
  token = Sys.getenv("SLACK_API_TOKEN"),
  ...,
  action = c("push", "pop"),
  rate_limit
)
}
\arguments{
\item{slack_method}{character, method to invoke, Default: 'chat.postMessage'}

\item{token}{Character. Your Slack API token.}

\item{...}{arguments to pass to chat methods}

\item{action}{action to apply on.exit to call stack, Default: c("push", "pop")}

\item{rate_limit}{Integer. The maximum number of calls to make to this
function per minute. If NULL, calls will not be rate-limited. If unset,
calls will be limited per the rate limit guides in the \href{https://api.slack.com/methods}{Slack api method documentation}, if known. Note: If a limit
is set for this method in any call in this session, that rate limit will be
respected until rate_limit is set to NULL. Default: None.}
}
\value{
httr response

A list with an additional class corresponding to \code{slack_method}.
}
\description{
Post/edit/delete messages or blocks to a channel
}
\details{
chat methods available to post to
\itemize{
\item chat.postMessage
\item chat.delete
\item chat.update
}

For a full list of chat methods see \href{https://api.slack.com/methods}{here}
}
\examples{
\dontshow{if (Sys.getenv('SLACK_API_TOKEN') != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
chat_slack(
  text = 'my message',
  channel = "general",
  token = "my_api_token"
)
\dontshow{\}) # examplesIf}
}
