% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{delete_post}
\alias{delete_post}
\title{Delete post from a channel}
\usage{
delete_post(channel, ts, token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{channel}{character, Channel ID}

\item{ts}{character, timestamp}

\item{token}{Character. Your Slack API token. Default: Sys.getenv("SLACK_API_TOKEN")}
}
\value{
A list with an additional class corresponding to \code{slack_method}.
}
\description{
Delete a post from a channel based on the timestamp (ts).
}
