% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{file_last}
\alias{file_last}
\alias{file_stack}
\title{Manage Files}
\usage{
file_last()

file_stack()
}
\description{
Query/Remove files created by slackreprex. Each file is logged
for the id in an internal package environment. Which
serves the purpose of querying or removing specific files.
\itemize{
\item file_stack: lists in decreasing order the files sent during the session.
\item last_file: retrieves the last file sent.
\item rm_last_file: deletes the last file from the specified channel.
}
}
\examples{
file_last()
file_stack()
}
