% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{parse_call}
\alias{parse_call}
\title{Parse Calls}
\usage{
parse_call()
}
\value{
api method to be used
}
\description{
Function used to translate R side functions to slack api methods
to interact with slack.
}
