% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{post_last}
\alias{post_last}
\alias{post_stack}
\title{Manage Posts}
\usage{
post_last()

post_stack()
}
\description{
Query/Remove posts created by slackreprex. Each post is logged
for channel and timestamp (ts) in an internal package environment. Which
serves the purpose of querying or removing specific posts.
\itemize{
\item post_stack: lists in decreasing order the posts sent during the session.
\item last_post: retrieves the last post sent.
}
}
