% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{post_slack}
\alias{post_slack}
\title{Call a Slack API Method}
\usage{
post_slack(
  slack_method,
  token = Sys.getenv("SLACK_API_TOKEN"),
  max_results = Inf,
  max_calls = Inf,
  limit = 1000L,
  ...,
  rate_limit
)
}
\arguments{
\item{slack_method}{Character. The Slack API method to call, such as
"conversations.history".}

\item{token}{Character. Your Slack API token.}

\item{max_results}{Integer. The maximum number of results to return (total).
Note: The actual maximum will be \code{max_results} rounded up to a
multiple of \code{limit}. Default: Inf}

\item{max_calls}{Integer. The maximum number of separate API calls to make
while constructing the response. Default: Inf}

\item{limit}{Integer. The number of results to fetch per call. Default: 1000L}

\item{...}{Additional arguments to pass to the body of the method call. Note:
if you pass an explicit \code{limit}, this may cause conflicts. We
recommend using \code{max_results} and \code{max_calls}.}

\item{rate_limit}{Integer. The maximum number of calls to make to this
function per minute. If NULL, calls will not be rate-limited. If unset,
calls will be limited per the rate limit guides in the \href{https://api.slack.com/methods}{Slack api method documentation}, if known. Note: If a limit
is set for this method in any call in this session, that rate limit will be
respected until rate_limit is set to NULL. Default: None.}
}
\value{
A list with an additional class corresponding to \code{slack_method}.
}
\description{
Safely call a Slack API method, with automatic pagination.
}
\examples{
\dontshow{if (Sys.getenv("SLACK_API_TOKEN") != "") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
post_slack(
  slack_method = "conversations.history",
  channel = "general",
  token = "my_api_token"
)
\dontshow{\}) # examplesIf}
}
