% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{slack_err}
\alias{slack_err}
\title{Slack API Error}
\usage{
slack_err(obj)
}
\arguments{
\item{obj}{httr response content}
}
\value{
character
}
\description{
Returns error message from a Slack API response
}
