% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{upload_slack}
\alias{upload_slack}
\title{Upload file or Content to Slack}
\usage{
upload_slack(slack_method = "files.upload", ..., token)
}
\arguments{
\item{slack_method}{Character. The Slack API method to call, Default: 'file.upload'}

\item{...}{Additional arguments to pass to the body of the method call.}

\item{token}{Character. Your Slack API token.}
}
\value{
A list with an additional class corresponding to \code{slack_method}.
}
\description{
Safely upload files and snippets to Slack
}
\examples{
\dontrun{
upload_slack(
  channel = "general",
  token = "my_api_token",
  content = 'wow'
)

tf <- tempfile(fileext = '.r')
cat(
  utils::capture.output(utils::sessionInfo()),
  file = tf,
  sep = '\n'
)

upload_slack(
  channel = "general",
  token = "my_api_token",
  file = tf,
  filename = 'sessionInfo.R',
  filetype = 'r',
  initial_comment = 'here is my session info',
  title = 'R sessionInfo'
)

unlink(tf)

}
}
