% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{validate_response}
\alias{validate_response}
\title{Validate Slack API Response}
\usage{
validate_response(res)
}
\arguments{
\item{res}{httr response}
}
\value{
httr response content
}
\description{
Check Slack API Response for errors
}
