% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{chat_slack}
\alias{chat_slack}
\title{Interact with Chat API}
\usage{
chat_slack(
  slack_method = "chat.postMessage",
  token = Sys.getenv("SLACK_API_TOKEN"),
  ...,
  action = c("push", "pop")
)
}
\arguments{
\item{slack_method}{character, method to invoke, Default: 'chat.postMessage'}

\item{token}{Character. Your Slack API token.}

\item{...}{arguments to pass to chat methods}

\item{action}{action to apply on.exit to call stack, Default: c("push", "pop")}
}
\value{
httr response
}
\description{
Post/edit/delete messages or blocks to a channel
}
\details{
chat methods available to post to
\itemize{
\item chat.postMessage
\item chat.delete
\item chat.update
For a full list of chat methods see \href{https://api.slack.com/methods}{here}
}
}
\examples{
\dontrun{
post_slack(
  text = 'my message',
  channel = "general",
  token = "my_api_token"
)
}
}
