% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{files_slack}
\alias{files_slack}
\title{Interact with File API}
\usage{
files_slack(
  slack_method = "files.upload",
  ...,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{slack_method}{character, method to invoke, Default: 'files.upload'}

\item{...}{arguments to pass to chat methods}

\item{token}{Character. Your Slack API token.}
}
\value{
httr response
}
\description{
upload/info/list/delete files
}
\details{
chat methods available to post to
\itemize{
\item files.upload
\item files.delete
\item files.info
\item files.list
For a full list of chat methods see \href{https://api.slack.com/methods}{here}
}
}
\examples{
\dontrun{
files_slack(
  method = 'files.upload',
  channel = "general",
  token = Sys.getenv('SLACK_API_TOKEN'),
  content = 'wow'
)

tf <- tempfile(fileext = '.r')
cat(
  utils::capture.output(utils::sessionInfo()),
  file = tf,
  sep = '\n'
)

files_slack(
  method = 'files.upload',
  channel = "general",
  token = Sys.getenv('SLACK_API_TOKEN'),
  file = tf,
  filename = 'sessionInfo.R',
  filetype = 'r',
  initial_comment = 'here is my session info',
  title = 'R sessionInfo'
)

unlink(tf)

}
}
