% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{post_slack}
\alias{post_slack}
\title{FUNCTION_TITLE}
\usage{
post_slack(slack_method, max_results = Inf, max_calls = Inf,
  paginate = TRUE, ...)
}
\arguments{
\item{slack_method}{PARAM_DESCRIPTION}

\item{max_results}{PARAM_DESCRIPTION, Default: Inf}

\item{max_calls}{PARAM_DESCRIPTION, Default: Inf}

\item{paginate}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
