% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{slack_err}
\alias{slack_err}
\title{Slack API Error}
\usage{
slack_err(obj)
}
\arguments{
\item{obj}{httr response content}
}
\value{
character
}
\description{
Returns error message from a Slack API response
}
\examples{
# obj represents an object returned from a post to the 'Slack' API.
obj <- list(
  error = "Example error message.",
  response_metadata = list(messages = c("Example messages.", "Another."))
)
slack_err(obj)
}
